+++ packages/vseed/src/pipeline/spec/pipes/vchart/axes/xLinear.ts
@@ -39,14 +39,17 @@ export const xLinear: SpecPipe = (spec, context) => {
     inverse,
     max,
     min,
+    log,
+    logBase = 10,
   } = config
 
   const linearAxis = {
     visible,
-    type: 'linear',
+    type: log ? 'log' : 'linear',
+    base: logBase,
     orient: 'bottom',
     nice,
-    zero,
+    zero: log ? false : zero,
     inverse,
     max,
     min,