+++ packages/vseed/src/pipeline/utils/format/createNumFormatter.ts
@@ -1,17 +1,45 @@
+import { isNumber } from 'remeda'
 import { intl } from '../../../i18n'
 import type { Formatter, Locale, NumFormat } from 'src/types'
 
-export const createNumFormatter = (format: NumFormat): Formatter => {
+export const createNumFormatter = (format: NumFormat, locale: Locale = intl.getLocale()): Formatter => {
   const {
     type = 'number',
     ratio = 1,
     symbol = '',
     thousandSeparator = true,
-    decimalPlaces = 2,
-    round = 'round',
     prefix = '',
     suffix = '',
+
+    fractionDigits = 2,
+    significantDigits,
+    roundingMode = 'halfExpand',
+    roundingPriority = 'auto',
   } = format || {}
+
+  const numFormatterOptions: Intl.NumberFormatOptions = {
+    style: 'decimal',
+    notation: type === 'scientific' ? 'scientific' : 'standard',
+    // @ts-expect-error roundingMode is not in NumberFormatOptions
+    roundingMode,
+    roundingPriority,
+  }
+  console.log('debug roundingMode', roundingMode)
+
+  if (isNumber(fractionDigits)) {
+    if (fractionDigits >= 0) {
+      numFormatterOptions.minimumFractionDigits = fractionDigits
+      numFormatterOptions.maximumFractionDigits = fractionDigits
+    }
+  }
+  if (isNumber(significantDigits) && significantDigits > 0) {
+    if (significantDigits > 0) {
+      numFormatterOptions.minimumSignificantDigits = significantDigits
+      numFormatterOptions.maximumSignificantDigits = significantDigits
+    }
+  }
+
+  const numFormatter = new Intl.NumberFormat(locale, numFormatterOptions)
 
   return (value?: number | string) => {
     let num = Number(value)
@@ -31,11 +59,8 @@ export const createNumFormatter = (format: NumFormat): Formatter => {
       num = num / (ratio || 1)
     }
 
-    // rounding
-    const multiplier = 10 ** decimalPlaces
-    num = Math[round](num * multiplier) / multiplier
-
-    let numStr = num.toFixed(decimalPlaces)
+    // format value
+    let numStr = numFormatter.format(num)
 
     // add thousand separator
     if (thousandSeparator) {
@@ -54,54 +79,14 @@ export const autoNumFormatter = (value?: number | string, locale: Locale = intl.
   const num = Number(value)
   if (Number.isNaN(num)) return String(value)
 
-  const countDecimalPlaces = (num: number) => {
-    if (Number.isInteger(num)) return 0
-
-    const str = num.toString()
-    if (str.indexOf('e-') > -1) {
-      return parseInt(str.split('e-')[1])
-    }
-
-    const decimalPart = str.split('.')[1]
-    if (!decimalPart) return 0
-
-    const decimalPlaces = decimalPart.replace(/0+$/, '').length
-    return Math.max(2, decimalPlaces)
+  const numFormatterOptions: Intl.NumberFormatOptions = {
+    style: 'decimal',
+    notation: 'compact',
   }
+  numFormatterOptions.minimumFractionDigits = 0
+  numFormatterOptions.maximumFractionDigits = 2
 
-  const numFormat: NumFormat = {
-    type: 'number',
-    decimalPlaces: countDecimalPlaces(num),
-    round: 'round',
-    thousandSeparator: true,
-  }
+  const numFormatter = new Intl.NumberFormat(locale, { ...numFormatterOptions })
 
-  const rules = NUMBER_FORMAT_RULES[locale] || NUMBER_FORMAT_RULES['default']
-
-  for (const rule of rules) {
-    if (num >= rule.threshold) {
-      numFormat.ratio = rule.ratio
-      numFormat.symbol = rule.symbol
-      break // 使用第一个匹配的规则
-    }
-  }
-
-  return createNumFormatter(numFormat)(value)
-}
-
-const NUMBER_FORMAT_RULES = {
-  'zh-CN': [
-    { threshold: 100000000, ratio: 100000000, symbol: '亿' },
-    { threshold: 10000, ratio: 10000, symbol: '万' },
-  ],
-  'en-US': [
-    { threshold: 1000000000, ratio: 1000000000, symbol: 'B' },
-    { threshold: 1000000, ratio: 1000000, symbol: 'M' },
-    { threshold: 1000, ratio: 1000, symbol: 'K' },
-  ],
-  default: [
-    { threshold: 1000000000, ratio: 1000000000, symbol: 'B' },
-    { threshold: 1000000, ratio: 1000000, symbol: 'M' },
-    { threshold: 1000, ratio: 1000, symbol: 'K' },
-  ],
+  return numFormatter.format(Number(value))
 }