+++ apps/website/CHANGELOG.md
@@ -1,5 +1,12 @@
 # website
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @visactor/vseed@0.0.19
+
 ## 0.0.18
 
 ### Patch Changes